Adjust styling and layout
Nothing super major but we move to simpler fonts which are more
bootstrap in feel and no longer float the article details in a weird way
to the right. And get rid of the categories at the top as I'm not really
using them.

Some other subtle changes. More tweaks needed I imagine but we're
edging in a good direction.